+++ web/src/components/Button.tsx
@@ -1,19 +1,16 @@
 import React from 'react';
 import { twMerge } from 'tailwind-merge';
 
-// This value is selected based on appropriate size given most labels in most languages.
-const BUTTON_MIN_WIDTH_CLASS = 'min-w-[232px]';
-
-// TODO: This should extend either HTMLButtonElement or HTMLAnchorElement, but I could not get it
-// to properly select the correct one based on given props.
 interface ButtonProps {
   icon?: React.ReactNode;
   children?: React.ReactNode;
   disabled?: boolean;
-  background?: string;
-  textColor?: string;
+  size: 'sm' | 'lg' | 'xl';
+  type: 'primary' | 'secondary' | 'secondary-elevated' | 'tertiary' | 'link';
   href?: string;
-  className?: string;
+  backgroundClasses?: string;
+  forgroundClasses?: string;
+  focusOutlineColor?: string;
   onClick?: () => void;
 }
 
@@ -21,34 +18,133 @@ export function Button({
   icon,
   children,
   disabled,
-  background,
-  textColor,
   href,
-  className,
+  backgroundClasses, // backgroundColor, borderColor, margin, etc.
+  forgroundClasses, // textColor, etc.
+  focusOutlineColor,
+  size,
+  type,
   onClick,
 }: ButtonProps) {
   const renderAsLink = Boolean(href);
   const As = renderAsLink ? 'a' : 'button';
-  const type = renderAsLink ? undefined : 'button';
+  const componentType = renderAsLink ? undefined : 'button';
   const isIconOnly = !children && Boolean(icon);
 
   return (
-    <As
+    <div
       className={twMerge(
-        `mx-2 my-3 flex min-h-[45px] w-fit items-center justify-center gap-x-2 rounded-full bg-white px-2 py-3 text-md font-bold shadow-[0px_0px_13px_rgb(0_0_0/12%)] transition duration-200 hover:shadow-[0px_0px_23px_rgb(0_0_0/20%)] dark:bg-gray-600 dark:hover:shadow-[0px_0px_23px_rgb(0_0_0/50%)]`,
-        !isIconOnly && BUTTON_MIN_WIDTH_CLASS,
-        `${disabled ? 'opacity-60 hover:shadow-[0px_0px_13px_rgb(0_0_0/12%)]' : ''}`,
-        className
+        `items-center justify-center rounded-full ${getBackground(type, disabled)}`,
+        backgroundClasses
       )}
-      translate="no"
-      disabled={disabled}
-      style={{ color: textColor, background: background }}
-      href={href}
-      type={type}
-      onClick={onClick}
     >
-      {icon}
-      {children}
-    </As>
+      <As
+        className={twMerge(
+          `flex w-full flex-row items-center justify-center rounded-full text-sm font-semibold disabled:text-neutral-400
+        disabled:hover:bg-inherit disabled:dark:text-gray-500 ${getSize(
+          size,
+          type,
+          isIconOnly
+        )}
+        ${getFocused(focusOutlineColor)} ${getForeground(type)} ${getHover(type)}`,
+          forgroundClasses
+        )}
+        disabled={disabled}
+        href={href}
+        type={componentType}
+        onClick={onClick}
+      >
+        {icon}
+        {children}
+      </As>
+    </div>
   );
+}
+
+function getHover(type: string) {
+  switch (type) {
+    case 'primary': {
+      return 'hover:bg-black/20';
+    }
+    default: {
+      return 'hover:bg-neutral-400/10';
+    }
+  }
+}
+
+function getBackground(type: string, disabled: boolean | undefined) {
+  switch (type) {
+    case 'primary': {
+      if (disabled) {
+        return 'bg-zinc-50 dark:bg-gray-800 border border-neutral-200 dark:border-gray-700';
+      }
+      return 'bg-brand-green';
+    }
+    case 'secondary': {
+      return 'border dark:border-gray-700 border-neutral-200 bg-inherit';
+    }
+    case 'secondary-elevated': {
+      return 'border dark:border-gray-700 border-neutral-200 bg-neutral-100 dark:bg-gray-800';
+    }
+    default: {
+      return 'bg-inherit';
+    }
+  }
+}
+
+function getFocused(focusOutlineColor: string | undefined) {
+  const outline =
+    'focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 ';
+  return focusOutlineColor
+    ? outline + focusOutlineColor
+    : outline + 'focus-visible:outline-brand-green';
+}
+
+function getForeground(type: string) {
+  switch (type) {
+    case 'primary': {
+      return 'text-white';
+    }
+    case 'link': {
+      return 'text-emerald-800 dark:text-emerald-500';
+    }
+    default: {
+      return 'text-black dark:text-white';
+    }
+  }
+}
+
+function getSize(size: string, type: string, isIconOnly: boolean) {
+  if (isIconOnly) {
+    switch (size) {
+      case 'sm': {
+        return 'min-w-7 min-h-7';
+      }
+      case 'lg': {
+        return 'min-w-11 min-h-11';
+      }
+      case 'xl': {
+        return 'min-w-14 min-h-14';
+      }
+    }
+  }
+
+  switch (size) {
+    case 'sm': {
+      return 'min-w-6 min-h-6 px-2 py-1 gap-x-1 text-sm';
+    }
+    case 'lg': {
+      return type == 'link'
+        ? 'px-4 py-2 gap-x-2 text-sm'
+        : 'min-w-10 min-h-10 px-6 py-3 gap-x-1.5 text-sm';
+    }
+    case 'xl': {
+      return type == 'link'
+        ? 'px-4 py-2 gap-x-2 text-base'
+        : 'min-w-12 min-h-12 px-8 py-4 gap-x-1.5 text-base';
+    }
+    default: {
+      return '';
+    }
+  }
 }
